[Contract Payment] - * replace contract payment request with contract payment
